requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Django==2.1.1
django-redis==4.9.0
djangorestframework==3.10.3
django-mysql==3.2.0
drf-extensions==0.4.0
google-api-core==1.14.3
google-auth==1.7.0
google-cloud-tasks==0.4.0
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.11.4
grpcio==1.24.3
idna==2.8
mysqlclient==1.4.5
protobuf==3.10.0
psycopg2-binary==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
pytz==2019.3
redis==2.10.6
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.6
